requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
line-bot-sdk==3.5.0
gspread==6.1.2
oauth2client==4.1.3
google-auth==2.30.0
openai==2.26.0
httpx==0.28.1
pydantic==2.7.4
python-dotenv==1.0.1
